Refactor combo rule classes and fix naming
Migrated related code from AbstractCellRuleCombo to new AbstractRuleCombo base class. Created AbstractAggregateRuleCombo to better handle aggregate rules. Adapted ComboSum and other classes to follow this structure. Renamed typographical error 'Abstarct' to 'Abstract' in class names. Various changes have been made to ensure relevant classes return 'float' type. README has been updated to reflect this refactor and improvements.
