+++ src/Rules/Aggregate/AbstarctAggregateRule.php
@@ -31,7 +31,7 @@ abstract class AbstarctAggregateRule extends AbstarctRule
      */
     abstract public function validateRule(array &$columnValues): ?string;
 
-    protected function getRuleCode(): string
+    protected function getRuleCode(?string $mode = null): string
     {
         return 'ag:' . parent::getRuleCode();
     }
