+++ src/Rules/Cell/ComboLength.php
@@ -22,12 +22,12 @@ final class ComboLength extends AbstractCellRuleCombo
 
     protected const HELP_TOP = ['Checks length of a string including spaces (multibyte safe).'];
 
-    protected function getExpected(): float|int|string
+    protected function getExpected(): float
     {
         return $this->getOptionAsInt();
     }
 
-    protected function getCurrent(string $cellValue): float|int|string
+    protected function getCurrent(string $cellValue): float
     {
         return \mb_strlen($cellValue);
     }
