+++ src/Rules/Cell/ComboWordCount.php
@@ -26,12 +26,12 @@ final class ComboWordCount extends AbstractCellRuleCombo
         'Example: "Hello World, 123" - 2 words only (123 is not a word).',
     ];
 
-    protected function getExpected(): float|int|string
+    protected function getExpected(): float
     {
         return $this->getOptionAsInt();
     }
 
-    protected function getCurrent(string $cellValue): float|int|string
+    protected function getCurrent(string $cellValue): float
     {
         // @phan-suppress-next-line PhanPartialTypeMismatchReturn
         return \str_word_count($cellValue, 0);
